remove movies and its controller, services, repo move parsing xml and mapping (item->episode) from service in quartz scheduler refactor episode repo

+++ src/main/java/io/skai/okta/internshipstreamingserviceparserxml/dto/Episode.java
@@ -13,7 +13,7 @@ public class Episode {
     long id;
     int season;
     int episodeNumber;
-    double imdbRating;
+    float imdbRating;
 
     String tvSeriesTitle;
